refactor(cloud): split control plane store domains
