+++ packages/cloud-server/src/postgres-domains/billing.ts
@@ -1,4 +1,4 @@
-import type { BillingSubscriptionRecord, UsageEventRecord } from '../control-plane-store.ts'
+import type { ArtifactUploadReservationRecord, BillingSubscriptionRecord, UsageEventRecord } from '../control-plane-store.ts'
 import { iso, isoOrNull, jsonRecord, numberValue, stringOrNull, type QueryRow } from './shared.ts'
 
 export function usageEventFromRow(row: QueryRow): UsageEventRecord {
@@ -27,6 +27,28 @@ export function billingSubscriptionFromRow(row: QueryRow): BillingSubscriptionRe
     currentPeriodEnd: isoOrNull(row.current_period_end),
     cancelAtPeriodEnd: row.cancel_at_period_end === true,
     metadata: jsonRecord(row.metadata),
+    createdAt: iso(row.created_at),
+    updatedAt: iso(row.updated_at),
+  }
+}
+
+export function artifactUploadReservationFromRow(row: QueryRow): ArtifactUploadReservationRecord {
+  return {
+    orgId: String(row.org_id),
+    tenantId: String(row.tenant_id),
+    userId: String(row.user_id),
+    sessionId: String(row.session_id),
+    artifactId: String(row.artifact_id),
+    objectKey: String(row.object_key),
+    filename: String(row.filename),
+    contentType: stringOrNull(row.content_type),
+    quotaKey: stringOrNull(row.quota_key),
+    quotaWindowMs: row.quota_window_ms === null || row.quota_window_ms === undefined ? null : numberValue(row.quota_window_ms),
+    quotaWindowStartedAtMs: row.quota_window_started_at_ms === null || row.quota_window_started_at_ms === undefined ? null : numberValue(row.quota_window_started_at_ms),
+    reservedBytes: numberValue(row.reserved_bytes),
+    settledBytes: row.settled_bytes === null || row.settled_bytes === undefined ? null : numberValue(row.settled_bytes),
+    status: String(row.status) as ArtifactUploadReservationRecord['status'],
+    expiresAt: iso(row.expires_at),
     createdAt: iso(row.created_at),
     updatedAt: iso(row.updated_at),
   }
